// frontend/src/api/generated/model/projectMember.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gantry Board API
 * AI Agent Orchestration Kanban Board API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberRole } from './memberRole';

export interface ProjectMember {
  created_at: string;
  project_id: string;
  role: MemberRole;
  user_email: string;
  user_id: string;
  user_name: string;
}